+++ test/dotnet-watch.Tests/TestUtilities/TestRuntimeProcessLauncher.cs
@@ -1,8 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Microsoft.Build.Graph;
-
 namespace Microsoft.DotNet.Watch.UnitTests;
 
 internal class TestRuntimeProcessLauncher(ProjectLauncher projectLauncher) : IRuntimeProcessLauncher
@@ -32,5 +30,38 @@ internal class TestRuntimeProcessLauncher(ProjectLauncher projectLauncher) : IRu
     {
         TerminateLaunchedProcessesImpl?.Invoke();
         return ValueTask.CompletedTask;
+    }
+
+    public async Task<RunningProject> Launch(string projectPath, string workingDirectory, CancellationToken cancellationToken)
+    {
+        var projectOptions = new ProjectOptions()
+        {
+            IsMainProject = false,
+            Representation = new ProjectRepresentation(projectPath, entryPointFilePath: null),
+            WorkingDirectory = workingDirectory,
+            Command = "run",
+            CommandArguments = ["--project", projectPath],
+            LaunchEnvironmentVariables = [],
+            LaunchProfileName = default,
+        };
+
+        RestartOperation? startOp = null;
+        startOp = new RestartOperation(async cancellationToken =>
+        {
+            var result = await ProjectLauncher.TryLaunchProcessAsync(
+                projectOptions,
+                onOutput: null,
+                onExit: null,
+                restartOperation: startOp!,
+                cancellationToken);
+
+            Assert.NotNull(result);
+
+            await result.Clients.WaitForConnectionEstablishedAsync(cancellationToken);
+
+            return result;
+        });
+
+        return await startOp(cancellationToken);
     }
 }
